requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
Pillow==10.0.1
plotly==5.18.0
scikit_learn==1.2.2
seaborn==0.13.2
streamlit==1.31.0
streamlit_option_menu==0.3.12
tensorflow
